pyproject.toml
[project]
name = "AmazonWarriors"
version = "0.1.0"
description = "An ArcadeActions demo with python-statemachine"
authors = [{name = "Brandon Corfman", email = "h9tbgnbbk@privaterelay.appleid.com"}]
license = {file="LICENSE"}
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    "arcade-actions>=0.8.2",
]

[project.optional-dependencies]
statemachine = [
    "python-statemachine",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[tool.uv.sources]
python-statemachine = { git = "https://github.com/bcorfman/python-statemachine", rev = "develop" }

[tool.pytest.ini_options]
addopts = "--cov-branch --cov-report term --cov=amazonwarriors"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers=["unit: Unit tests.",
         "integration: Integration tests."]
         
[tool.coverage.run]
source = ["amazonwarriors"]

[tool.ruff]
line-length = 120
fix = true
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
extend-fixable = ["E", "F", "UP", "B", "SIM", "I"]
